wp-now: rename executePHPFile to executePHP
