+++ packages/wp-now/src/tests/execute-php.spec.ts
@@ -1,9 +1,9 @@
 import fs from 'fs-extra';
 import path from 'path';
-import { executePHPFile } from '../execute-php-file';
+import { executePHP } from '../execute-php';
 import getWpNowConfig from '../config';
 
-const exampleDir = path.join(__dirname, 'execute-php-file');
+const exampleDir = path.join(__dirname, 'execute-php');
 
 test('php file execution in index mode', async () => {
 	const resultFilePath = path.join(exampleDir, 'hello-world-result.txt');
@@ -12,7 +12,7 @@ test('php file execution in index mode', async () => {
 	const options = await getWpNowConfig({
 		path: exampleDir,
 	});
-	await executePHPFile(
+	await executePHP(
 		['php', path.join(exampleDir, 'hello-world.php')],
 		options
 	);
@@ -25,21 +25,21 @@ test('php file execution for each PHP Version', async () => {
 	const options = await getWpNowConfig({
 		path: exampleDir,
 	});
-	await executePHPFile(['php', path.join(exampleDir, 'php-version.php')], {
+	await executePHP(['php', path.join(exampleDir, 'php-version.php')], {
 		...options,
 		phpVersion: '7.4',
 	});
 	let output = fs.readFileSync(resultFilePath, 'utf8');
 	expect(output.substring(0, 16)).toBe('PHP Version: 7.4');
 
-	await executePHPFile(['php', path.join(exampleDir, 'php-version.php')], {
+	await executePHP(['php', path.join(exampleDir, 'php-version.php')], {
 		...options,
 		phpVersion: '8.0',
 	});
 	output = fs.readFileSync(resultFilePath, 'utf8');
 	expect(output.substring(0, 16)).toBe('PHP Version: 8.0');
 
-	await executePHPFile(['php', path.join(exampleDir, 'php-version.php')], {
+	await executePHP(['php', path.join(exampleDir, 'php-version.php')], {
 		...options,
 		phpVersion: '8.2',
 	});
